feat(backend): delete blog only on authorize users

+++ backend/controllers/blogs.js
@@ -40,6 +40,20 @@ blogsRouter.post("/", async (request, response) => {
 
 blogsRouter.delete("/:id", async (request, response) => {
   const id = request.params.id;
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(id);
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response
+      .status(401)
+      .json({ error: "You are not authorize to delete this blog" });
+  }
+
   await Blog.findOneAndDelete(id);
   response.status(204).end();
 });
